Seeding problems
Currently having trouble with the seeded tickets duplicating themselves.  Seems like it is happening when I log in and out as regular users.

+++ DragonBugs2020/Controllers/ProjectsController.cs
@@ -44,27 +44,34 @@ namespace DragonBugs2020.Controllers
             else if (User.IsInRole("ProjectManager"))
             {
                 var projectIds = new List<int>();
+                model = new List<Project>();
                 //grab data that I need here
                 var userProjects = _context.ProjectUsers.Where(pu => pu.UserId == userId).ToList();
 
                 foreach (var record in userProjects)
                 {
-                    projectIds.Add(record.ProjectId);
+                    projectIds.Add(_context.Projects.Find(record.ProjectId).Id);
                 }
                 foreach (var id in projectIds)
                 {
-                    var project = _context.Projects.Find(id);
-                    model.Add(project);
+                    var project = _context.Projects.Where(t => t.Id == id).ToList();
+                    model.AddRange(project);
                 }
+
             }
 
             else if (User.IsInRole("Developer"))
             {
                 model = await _btProjectService.ListUserProjects(userId);
-                
+
             }
 
             else if (User.IsInRole("Submitter"))
+            {
+                model = await _btProjectService.ListUserProjects(userId);
+
+            }
+            else if (User.IsInRole("NewUser"))
             {
                 model = await _btProjectService.ListUserProjects(userId);
 
@@ -87,6 +94,17 @@ namespace DragonBugs2020.Controllers
                 .Include(p => p.ProjectUsers)
                 .ThenInclude(p => p.User)
                 .FirstOrDefaultAsync(m => m.Id == id);
+
+            project.Tickets = await _context.Tickets
+                .Where(t => t.ProjectId == id)
+                .Include(t => t.DeveloperUser)
+                .Include(t => t.OwnerUser)
+                .Include(t => t.Project)
+                .Include(t => t.TicketPriority)
+                .Include(t => t.TicketStatus)
+                .Include(t => t.TicketType)
+                .ToListAsync();
+
             if (project == null)
             {
                 return NotFound();
@@ -110,13 +128,26 @@ namespace DragonBugs2020.Controllers
         [Authorize(Roles = "Admin, ProjectManager")]
         public async Task<IActionResult> Create([Bind("Id,Name,ImagePath,ImageData")] Project project)
         {
-            if (ModelState.IsValid)
+            if (!User.IsInRole("Demo"))
             {
-                _context.Add(project);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(project);
+                    await _context.SaveChangesAsync();
+                    ProjectUser record = new ProjectUser { UserId = _userManager.GetUserId(User), ProjectId = project.Id };
+                    _context.ProjectUsers.Add(record);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(MyProjects));
+                }
+
+                return View(project);
             }
-            return View(project);
+            else
+            {
+                TempData["DemoLockout"] = "Your changes will not be saved.  To make changes to the database please log in as a full user.";
+
+                return RedirectToAction(nameof(MyProjects));
+            }
         }
 
         // GET: Projects/Edit/5
@@ -148,59 +179,31 @@ namespace DragonBugs2020.Controllers
             {
                 return NotFound();
             }
-
-            if (ModelState.IsValid)
+            if (!User.IsInRole("Demo"))
             {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(project);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!ProjectExists(project.Id))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(project);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!ProjectExists(project.Id))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction(nameof(Index));
                 }
                 return RedirectToAction(nameof(Index));
             }
             return View(project);
-        }
-
-        // GET: Projects/Delete/5
-        [Authorize(Roles = "Admin, ProjectManager")]
-        public async Task<IActionResult> Delete(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var project = await _context.Projects
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (project == null)
-            {
-                return NotFound();
-            }
-
-            return View(project);
-        }
-
-        // POST: Projects/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        [Authorize(Roles = "Admin, ProjectManager")]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var project = await _context.Projects.FindAsync(id);
-            _context.Projects.Remove(project);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
         }
 
         private bool ProjectExists(int id)
@@ -214,14 +217,18 @@ namespace DragonBugs2020.Controllers
         {
             var model = new ManageProjectUsersViewModel();
             var project = _context.Projects
+
                 .Include(p => p.ProjectUsers)
                 .ThenInclude(p => p.User)
                 .FirstAsync(p => p.Id == id);
-
-            model.Project = await project;
-            List<BTUser> users = await _context.Users.ToListAsync();
-            List<BTUser> members = (List<BTUser>)await _btProjectService.UsersOnProject(id);
-            model.Users = new MultiSelectList(users, "Id", "FullName", members);
+            if (!User.IsInRole("Demo"))
+            {
+                model.Project = await project;
+                List<BTUser> users = await _context.Users.ToListAsync();
+                List<BTUser> members = (List<BTUser>)await _btProjectService.UsersOnProject(id);
+                model.Users = new MultiSelectList(users, "Id", "FullName", members);
+                return View(model);
+            }
             return View(model);
         }
 
@@ -230,27 +237,33 @@ namespace DragonBugs2020.Controllers
         [Authorize(Roles = "Admin, ProjectManager")]
         public async Task<IActionResult> AssignUsers(ManageProjectUsersViewModel model)
         {
-            if (ModelState.IsValid)
+            if (!User.IsInRole("Demo"))
             {
-                if (model.SelectedUsers != null)
+                if (ModelState.IsValid)
                 {
-                    var currentMembers = await _context.Projects.Include(p => p.ProjectUsers).FirstOrDefaultAsync(p => p.Id == model.Project.Id);
-                    List<string> memberIds = currentMembers.ProjectUsers.Select(u => u.UserId).ToList();
-                    memberIds.ForEach(i => _btProjectService.AddUserToProject(i, model.Project.Id));
-                    foreach (string id in memberIds)
+                    if (model.SelectedUsers != null)
                     {
-                        await _btProjectService.RemoveUserFromProject(id, model.Project.Id);
+                        var currentMembers = await _context.Projects.Include(p => p.ProjectUsers).FirstOrDefaultAsync(p => p.Id == model.Project.Id);
+                        List<string> memberIds = currentMembers.ProjectUsers.Select(u => u.UserId).ToList();
+                        memberIds.ForEach(i => _btProjectService.AddUserToProject(i, model.Project.Id));
+                        foreach (string id in memberIds)
+                        {
+                            await _btProjectService.RemoveUserFromProject(id, model.Project.Id);
+                        }
+                        foreach (string id in model.SelectedUsers)
+                        {
+                            await _btProjectService.AddUserToProject(id, model.Project.Id);
+                        }
+                        return RedirectToAction("Details", "Projects", new { id = model.Project.Id });
                     }
-                    foreach (string id in model.SelectedUsers)
+                    else
                     {
-                        await _btProjectService.AddUserToProject(id, model.Project.Id);
+
                     }
-                    return RedirectToAction("Details", "Projects", new { id = model.Project.Id});
-                 }
-                else
-                {
-                    
+                    return View(model);
+
                 }
+
             }
             return View(model);
         }
@@ -262,10 +275,14 @@ namespace DragonBugs2020.Controllers
             var model = new ManageProjectUsersViewModel();
             var project = _context.Projects.Find(id);
 
-            model.Project = project;
-            List<BTUser> members = (List<BTUser>)await _btProjectService.UsersOnProject(id);
-            model.Users = new MultiSelectList(members, "Id", "FullName");
-            return View(model);
+            
+            
+                model.Project = project;
+                List<BTUser> members = (List<BTUser>)await _btProjectService.UsersOnProject(id);
+                model.Users = new MultiSelectList(members, "Id", "FullName");
+                return View(model);
+            
+            
         }
 
 
@@ -274,28 +291,26 @@ namespace DragonBugs2020.Controllers
         [Authorize(Roles = "Admin, ProjectManager")]
         public async Task<IActionResult> RemoveUsers(ManageProjectUsersViewModel model)
         {
-            if (ModelState.IsValid)
+            if (!User.IsInRole("Demo"))
             {
-                //var project = db.Projects.Find(model.projectId);
-                //project.Users.Clear();
-                if (model.SelectedUsers != null)
+                if (ModelState.IsValid)
                 {
-                    foreach (string id in model.SelectedUsers)
-                    {
-                        await _btProjectService.RemoveUserFromProject(id, model.Project.Id);
-                    }
-                    //return RedirectToAction("Index", "Projects");
 
-                    return RedirectToAction("Details", "Projects", new { id = model.Project.Id });
-                }
-                else
-                {
-                    //Send error message back
+                    if (model.SelectedUsers != null)
+                    {
+                        foreach (string id in model.SelectedUsers)
+                        {
+                            await _btProjectService.RemoveUserFromProject(id, model.Project.Id);
+                        }
+                        return RedirectToAction("Details", "Projects", new { id = model.Project.Id });
+                    }
+                    else
+                    {
+                        return View(model);
+                    }
                 }
             }
-            return View(model);
+            return RedirectToAction("Details", "Projects", new { id = model.Project.Id });
         }
-
-
     }
 }
